Use unstable channel as overlay.
This commit stops using `packageOverrides` to root the unstable
channel at `pkgs.unstable`, instead defining a simple overlay that adds
the attribute. This is better because it works through the compatability
overlay that allows Nix tools (eg. `nix-shell`) to use the overlays
defined in `nixpkgs.overlays` - previously, a package defined in a
overlay that depended on `pkgs.unstable` existing wouldn't work in
`nix-shell`.